merubah tulisan angka menjadi format rupiah

+++ lib/schema/chinesee_food_item.dart
@@ -6,7 +6,7 @@ part 'chinesee_food_item.g.dart';
 class ChineseeFoodItem {
   late Id id;
   late String title;
-  late double price;
+  late int price;
   late int quantity;
   late double userLat;
   late double userLong;
